refactor: create new `label` directory and separate modules and rename `Label` into `SemanticLabel`
