+++ perception_eval/test/util/dummy_object.py
@@ -17,7 +17,7 @@ from typing import List
 from typing import Tuple
 
 from perception_eval.common.label import AutowareLabel
-from perception_eval.common.label import Label
+from perception_eval.common.label import SemanticLabel
 from perception_eval.common.schema import FrameID
 from perception_eval.common.shape import Shape
 from perception_eval.common.shape import ShapeType
@@ -52,7 +52,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.5, 1.5, 1.5)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16) if use_unique_id else "0",
         ),
@@ -63,7 +63,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(0.5, 0.5, 0.5)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.BICYCLE, "bicycle", []),
+            semantic_label=SemanticLabel(AutowareLabel.BICYCLE, "bicycle", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16) if use_unique_id else "0",
         ),
@@ -74,7 +74,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.0, 1.0, 1.0)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16) if use_unique_id else "1",
         ),
@@ -87,7 +87,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.0, 1.0, 1.0)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16),
             pointcloud_num=10,
@@ -99,7 +99,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.0, 1.0, 1.0)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.BICYCLE, "bicycle", []),
+            semantic_label=SemanticLabel(AutowareLabel.BICYCLE, "bicycle", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16),
             pointcloud_num=10,
@@ -111,7 +111,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.0, 1.0, 1.0)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.PEDESTRIAN, "pedestrian", []),
+            semantic_label=SemanticLabel(AutowareLabel.PEDESTRIAN, "pedestrian", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16),
             pointcloud_num=10,
@@ -123,7 +123,7 @@ def make_dummy_data(
             orientation=Quaternion([0.0, 0.0, 0.0, 1.0]),
             shape=Shape(shape_type=ShapeType.BOUNDING_BOX, size=(1.0, 1.0, 1.0)),
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.MOTORBIKE, "motorbike", []),
+            semantic_label=SemanticLabel(AutowareLabel.MOTORBIKE, "motorbike", []),
             velocity=(1.0, 1.0, 1.0),
             uuid=token_hex(16),
             pointcloud_num=10,
@@ -148,7 +148,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             roi=(100, 100, 200, 100) if use_roi else None,
             uuid=token_hex(16) if use_roi else "0",
         ),
@@ -156,7 +156,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.BICYCLE, "bicycle", []),
+            semantic_label=SemanticLabel(AutowareLabel.BICYCLE, "bicycle", []),
             roi=(0, 0, 50, 50) if use_roi else None,
             uuid=token_hex(16) if use_roi else "1",
         ),
@@ -164,7 +164,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=0.9,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             roi=(200, 200, 200, 100) if use_roi else None,
             uuid=token_hex(16) if use_roi else "2",
         ),
@@ -174,7 +174,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=1.0,
-            semantic_label=Label(AutowareLabel.CAR, "car", []),
+            semantic_label=SemanticLabel(AutowareLabel.CAR, "car", []),
             roi=(100, 100, 200, 100) if use_roi else None,
             uuid=token_hex(16) if use_roi else "0",
         ),
@@ -182,7 +182,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=1.0,
-            semantic_label=Label(AutowareLabel.BICYCLE, "bicycle", []),
+            semantic_label=SemanticLabel(AutowareLabel.BICYCLE, "bicycle", []),
             roi=(0, 0, 50, 50) if use_roi else None,
             uuid=token_hex(16) if use_roi else "1",
         ),
@@ -190,7 +190,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=1.0,
-            semantic_label=Label(AutowareLabel.PEDESTRIAN, "pedestrian", []),
+            semantic_label=SemanticLabel(AutowareLabel.PEDESTRIAN, "pedestrian", []),
             roi=(200, 200, 200, 100) if use_roi else None,
             uuid=token_hex(16) if use_roi else "2",
         ),
@@ -198,7 +198,7 @@ def make_dummy_data2d(use_roi: bool = True) -> Tuple[List[DynamicObject2D], List
             unix_time=100,
             frame_id=frame_id,
             semantic_score=1.0,
-            semantic_label=Label(AutowareLabel.MOTORBIKE, "motorbike", []),
+            semantic_label=SemanticLabel(AutowareLabel.MOTORBIKE, "motorbike", []),
             roi=(300, 100, 50, 50) if use_roi else None,
             uuid=token_hex(16) if use_roi else "3",
         ),
